Lots of work on improving various aspects of rendering (#6776) particularly pertaining the RenderResizableCuboid: - Rewrote RenderResizableCuboid to be more based off of how Mantle/Tinkers render things. No longer renders backfaces so not much will be seen if you decide to jump inside an evap tower or something but for now it will do - Fixed scale calculations sometimes not allowing for it to get to 100% causing things like full dynamic tanks to have a small gap at the top - Fixed lighting of fluids in multiblocks being a bit incorrect at times - Fix how fission reactor scales down the heated coolant rendering to greatly reduce the amount of z-fighting that takes place (some still happens especially at greater distances but it looks a lot better up close)

+++ src/main/java/mekanism/client/render/transmitter/RenderMechanicalPipe.java
@@ -31,8 +31,8 @@ import net.minecraftforge.fluids.FluidStack;
 public class RenderMechanicalPipe extends RenderTransmitterBase<TileEntityMechanicalPipe> {
 
     private static final int stages = 100;
-    private static final double height = 0.45;
-    private static final double offset = 0.015;
+    private static final float height = 0.45F;
+    private static final float offset = 0.015F;
     //Note: this is basically used as an enum map (Direction), but null key is possible, which EnumMap doesn't support. 6 is used for null side
     private static final Int2ObjectMap<FluidRenderMap<Int2ObjectMap<Model3D>>> cachedLiquids = new Int2ObjectArrayMap<>(7);
 
@@ -113,71 +113,71 @@ public class RenderMechanicalPipe extends RenderTransmitterBase<TileEntityMechan
             model.setSideRender(side, false);
             model.setSideRender(side.getOpposite(), false);
         }
-        double stageRatio = (stage / (double) stages) * height;
+        float stageRatio = (stage / (float) stages) * height;
         switch (sideOrdinal) {
             case 0:
-                model.minX = 0.5 - stageRatio / 2;
+                model.minX = 0.5F - stageRatio / 2;
                 model.minY = 0;
-                model.minZ = 0.5 - stageRatio / 2;
+                model.minZ = 0.5F - stageRatio / 2;
 
-                model.maxX = 0.5 + stageRatio / 2;
-                model.maxY = 0.25 + offset;
-                model.maxZ = 0.5 + stageRatio / 2;
+                model.maxX = 0.5F + stageRatio / 2;
+                model.maxY = 0.25F + offset;
+                model.maxZ = 0.5F + stageRatio / 2;
                 break;
             case 1:
-                model.minX = 0.5 - stageRatio / 2;
-                model.minY = 0.25 - offset + stageRatio;
-                model.minZ = 0.5 - stageRatio / 2;
+                model.minX = 0.5F - stageRatio / 2;
+                model.minY = 0.25F - offset + stageRatio;
+                model.minZ = 0.5F - stageRatio / 2;
 
-                model.maxX = 0.5 + stageRatio / 2;
+                model.maxX = 0.5F + stageRatio / 2;
                 model.maxY = 1;
-                model.maxZ = 0.5 + stageRatio / 2;
+                model.maxZ = 0.5F + stageRatio / 2;
                 break;
             case 2:
-                model.minX = 0.25 + offset;
-                model.minY = 0.25 + offset;
+                model.minX = 0.25F + offset;
+                model.minY = 0.25F + offset;
                 model.minZ = 0;
 
-                model.maxX = 0.75 - offset;
-                model.maxY = 0.25 + offset + stageRatio;
-                model.maxZ = 0.25 + offset;
+                model.maxX = 0.75F - offset;
+                model.maxY = 0.25F + offset + stageRatio;
+                model.maxZ = 0.25F + offset;
                 break;
             case 3:
-                model.minX = 0.25 + offset;
-                model.minY = 0.25 + offset;
-                model.minZ = 0.75 - offset;
+                model.minX = 0.25F + offset;
+                model.minY = 0.25F + offset;
+                model.minZ = 0.75F - offset;
 
-                model.maxX = 0.75 - offset;
-                model.maxY = 0.25 + offset + stageRatio;
+                model.maxX = 0.75F - offset;
+                model.maxY = 0.25F + offset + stageRatio;
                 model.maxZ = 1;
                 break;
             case 4:
                 model.minX = 0;
-                model.minY = 0.25 + offset;
-                model.minZ = 0.25 + offset;
+                model.minY = 0.25F + offset;
+                model.minZ = 0.25F + offset;
 
-                model.maxX = 0.25 + offset;
-                model.maxY = 0.25 + offset + stageRatio;
-                model.maxZ = 0.75 - offset;
+                model.maxX = 0.25F + offset;
+                model.maxY = 0.25F + offset + stageRatio;
+                model.maxZ = 0.75F - offset;
                 break;
             case 5:
-                model.minX = 0.75 - offset;
-                model.minY = 0.25 + offset;
-                model.minZ = 0.25 + offset;
+                model.minX = 0.75F - offset;
+                model.minY = 0.25F + offset;
+                model.minZ = 0.25F + offset;
 
                 model.maxX = 1;
-                model.maxY = 0.25 + offset + stageRatio;
-                model.maxZ = 0.75 - offset;
+                model.maxY = 0.25F + offset + stageRatio;
+                model.maxZ = 0.75F - offset;
                 break;
             case 6:
                 //Null side
-                model.minX = 0.25 + offset;
-                model.minY = 0.25 + offset;
-                model.minZ = 0.25 + offset;
+                model.minX = 0.25F + offset;
+                model.minY = 0.25F + offset;
+                model.minZ = 0.25F + offset;
 
-                model.maxX = 0.75 - offset;
-                model.maxY = 0.25 + offset + stageRatio;
-                model.maxZ = 0.75 - offset;
+                model.maxX = 0.75F - offset;
+                model.maxY = 0.25F + offset + stageRatio;
+                model.maxZ = 0.75F - offset;
                 break;
         }
         cachedFluids.computeIfAbsent(fluid, f -> new Int2ObjectOpenHashMap<>()).putIfAbsent(stage, model);
